libcamera: Add libusb dependency to meson
As MediaDeviceUSB is thought of as a wrapper around libusb, we will need
this library as a dependency in our project for future use.
Currently this is not exposed as a configuration option but instead the MediaDeviceUSB
class is enabled if libusb dev was found during build time.

Tested against libusb-1.0.0-dev on ubuntu

Signed-off-by: Sophie Friedrich <dev@flowerpot.me>