libcamera: usb: Add USB device enumeration through sysfs
`DeviceEnumeratorSysfs` has been extended to support enumeration of usb
devices found in `/dev/bus/usb/*/*`. This seems to be safe as  libusb
has employed strategy for device enumeration on Linux over the past
decade.

Although `MediaDeviceUSB` is already capable of opening the device node
and retrieving a libusb device instance from it, there's currently no
further process in this patch to retrieve the USB device description and
entering it into the libcamera match system.

TODO: initialization of libusb happens in device_enumerator::create. It
  is untested / unknown if using the default libusb context there is
  safe.

Signed-off-by: Sophie Friedrich <dev@flowerpot.me>