libcamera: enumeration: Generalize DeviceMatch
Currently `DeviceMatch` is very specialized to handle `MediaDevice`s, which
has strong dependencies to V4L2 internals.

This patch enables `DeviceMatch` to match using key/value pairs instead
of a simple list of strings and introduces a `DeviceMatchEntityInterface` to
abstract away the likes of `MediaEntity` in order to search for a
device.

Signed-off-by: Sophie Friedrich <dev@flowerpot.me>