enumeration: sysfs: usb: do not scan main bus devices
`/dev/bus/usb/BBB/DDD` not only contains device nodes for attached USB
clients, but also the responsible bus devices. These seem to be
consistently enumerated as dev `BBB/001`.

This change was necessary as opening the bus endpoints could lead to a
blocking `open` while the system was waking for the bus to wake-up
again. As such camera enumeration could be delayed by multiple seconds.

Signed-off-by: Sophie Friedrich <dev@flowerpot.me>