fix: pipeline handlers: Stop exponential explosive calls to createPipelineHandlers
Currently the method `createPipelineHandlers` registered itself to the
`devicesAdded` signal at the end of each call. As the Signal object
supports multiple non-unique listeners connected to it, the former
method would be called exponentially often with each new emitted signal
on `devicesAdded` (i.e. with udev plugging in a new camera)

By attaching the `createPipelineHandlers` function to `devicesAdded` during
the first call to it / init of the `CameraManager::Private` this effect
is prevented.

Signed-off-by: Sophie Friedrich <dev@flowerpot.me>