pipeline: MobirAir: test: Allow buffer init
This patch essentially allows qcam and such to start and wait for frames
by the connected camera. Needed for that were working initialization and
export of buffers, which can be used with mmap in the application.

This currently mostly relies on `tmpfile` in order to (hopefully in the
future) dump data into buffer.

Signed-off-by: Sophie Friedrich <dev@flowerpot.me>