pipeline: MobirAir: test: Proof that I control the image
This patch proves that I'm able to render into the FrameBuffer with an
image, which the application can read out and render. The commit should
render a circle (with a few overflows at the border) onto the resulting
image.

Using temporary files here actually does the job here.

Note that this is still a test commit, as the heavy load is on the
userspace. It needs to be discussed in the future on whether this is
acceptable.

Signed-off-by: Sophie Friedrich <dev@flowerpot.me>