pipeline: MobirAir: test: fix image control
I previously hard forced the allowed buffers to 1 in the validate()
method and forgot to set the frame buffer metadata.

Both of them resulted in problems with qcam, as it would only request
the first frame and be done with it.

To show moving pictures, the render of the circle has changed.