pipline: MobirAir: Basic USB support
This patch introduces the MobirAirUSBWrapper class, which supports basic
USB initialization and communication with the MobirAir device. For now
this is a rough cut that works well enough. I'm not well versed enough
in C++ to write clean code and I'm still figuring stuff out as I'm
going.

The MobirAir essentially uses a protocol very similar to serial
terminals. It does not use any USB specific features like control
transfer or similar and uses two endpoints for the IN/OUT directions.
Additionally it overlays its own concept of commands on top of the USB
layer which model a request / response scheme.

The protocol only diverges from this after `StartX` has been called.
Afterwards it will send frames continuously until the execution of
`StopX`.

To accommodate that behaviour and still allow synchronous commands a
general USB communication thread has been introduced. Requests are
enqueued and will only be processed after the previous data transfers
have fully finished. This entails the guarantee that video frames will
be received fully before further requests are processed. This is
possible as the exact size of each frame and expected data buffer is
known beforehand.

To expose requests as an synchronous API to the outside, each request
contains a ConditionVariable, which the requesting thread will wait for
and that will be notified if the request has been fully processed (/
answered)

This patch shows the basic functionality by sending and completing a
request.

== Issues
- as mentioned above: this is only a rough draft which will require more
  tuning before it is in acceptable condition
- it is unclear so far how much should be abstracted away and how much
  is responsibility of the pipeline. There are a few things like the
  receive thread or the Request model, which other libusb based cameras
  could make use of - even if they don't use exactly the same endpoints
- the current patch is not very error proof, doesn't conform to the
  style guides nor does it do documentation in any way useful

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>