libcamera: Add libusb dependency to meson
As USBDevice is thought of as a wrapper around libusb, we will need this
library as a dependency in our project for future use. Currently this
this commit only includes setting up the libusb backend in the device
enumerator without further use.

This option is automatically enabled if libusb was found on the system.

Tested against libusb-1.0.0-dev on ubuntu

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>