libcamera: device_enumerator: Template create/remove device
The create(Media)Device and remove(Media)Device functions were
previously hardcoded to only work with the MediaDevice class. In order
to prepare for the new USBDevice class and remove code duplication
these two functions have have been worked  into a template function.

Note here that removeMediaDevice still exists as the removeDevice<T>
function is in the private scope of the DeviceEnumerator class.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>