libcamera: Fix non-abstract deconstructors for abstract class
The deconstructors in the abstract CameraDevice and DeviceMatch classes
were not abstract themself. This throws a compiler warning as it means
that during deconstruction of the abstract parent class, the actual
child class might not have their deconstructors be called. So its common
practice to do what this commit proposes.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>