libcamera: usb_device: Use devnode instead of vid/pid
Change USBDevice to work internally using a device node file, instead of
the vid/pid combination used previously. This is important, as a device
cannot be uniquely identified using just the vid/pid combination (all
USB devices which are the model have the same ids). Instead devices are
being mounted in Linux in the sysfs and all access runs through a
devnode.

This change also allows us to create a "populate" function which fills
the USB descriptors like vid/pid using libusb and as such make use of
the previously generalized `createDevice` function.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>