libcamera: device_enumerator: Improve USB udev implementation
The device enumeration through udev required serious work in order to
work on all USB devices and handle hotplugging correctly.

In this change we now observe the whole USB udev subsystem instead of
just input. Note here that we ignore all USB interfaces (e.g. hubs, root
controllers, …). They will presumably not be relevant for libcamera even
in the future, as every devices with an output stream (or similar) will
be enlisted on their own.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>